Initial reqSS
Source: Nate (rtcwMP)
I have excluded the quality argument, not needed atm. Can be added later, but tr_init needs to account for it.
Tested the command and it works locally.
Todo: finish the http side of things (see CL_RequestedSS in cl_controls.c) and make it silent on the client side.
Also, renamed flag operator for cvar restrict to shut the compiler up.