Update build-pk3-assets.sh
For the RtcwPro-Omnibots server, for now (until that server shuts down). There are many duplicated things that increase 1.2.9's rtcwpro_assets.pk3's filesize and haven't changed with even 1.2.7...

And that is why I propose going back to re-using the old PK3 and ONLY adding stuff to this newer one.