Additional features for demos, project sustaining upgrades
Add cg_notifyPlayerOnly to only show player perspective events in the notify text while playing a demo.

Begin removal of curl from the repository. Fetch and build the latest versions separately. Linux implementation in work.

Fixed a bug in WTV demos where the scoreboard would appear at the start of the demo due to openLimboMenu as a command in the demo.

Default the client to rtcwpro (in wolfMP.exe without a shortcut setting fs_game to rtcwpro)